Refactor: split monolithic main.ts into focused modules
Split the 1,327-line src/main.ts into three focused modules:
- src/holidays.ts: date logic, calendar math, holiday definitions, text formatting, phraseToMoment
- src/suggest.ts: DDSuggest editor suggestion class
- src/main.ts: slim plugin class, settings UI, and re-exports

Additional improvements:
- Extract duplicated month regex pattern (appeared 7x) into shared MONTH_PATTERN constant
- Pre-compile regex patterns at module level instead of on each function call
- Replace tsc+cp build with esbuild bundler (standard for Obsidian plugins)
- Remove 3 orphaned JS files (plugin.js, holidays.js, suggest.js) that were never loaded
- Fix zip script to only package main.js instead of unused modules
- Rewrite tests to use proper require() with mocked obsidian instead of brittle regex extraction
